+++ Models/Transaction.cs
@@ -10,7 +10,7 @@ namespace bank.Models
     public int TransactionId { get; set; }
     [Required]
     [RegularExpression(@"^\d+\.\d{0,2}$")]
-    public Decimal Amount { get; set; }
+    public Double Amount { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
     public int UserId { get; set; }
